Extract multiple modules from grid engine autoscaler library
