+++ workflows/pipe-common/pipeline/hpc/exec.py
@@ -1,0 +1,43 @@
+#  Copyright 2017-2023 EPAM Systems, Inc. (https://www.epam.com/)
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+import subprocess
+
+from pipeline.hpc.logger import Logger
+
+
+class ExecutionError(RuntimeError):
+    pass
+
+
+class CmdExecutor:
+
+    def __init__(self):
+        pass
+
+    def execute(self, command):
+        process = subprocess.Popen(command, shell=True, stderr=subprocess.PIPE, stdout=subprocess.PIPE)
+        out, err = process.communicate()
+        exit_code = process.wait()
+        if exit_code != 0:
+            exec_err_msg = 'Command \'%s\' execution has failed. Out: %s Err: %s.' % (command, out.rstrip(), err.rstrip())
+            Logger.warn(exec_err_msg)
+            raise ExecutionError(exec_err_msg)
+        return out
+
+    def execute_to_lines(self, command):
+        return self._non_empty(self.execute(command).splitlines())
+
+    def _non_empty(self, elements):
+        return [element for element in elements if element.strip()]
