+++ workflows/pipe-common/pipeline/hpc/logger.py
@@ -1,0 +1,34 @@
+#  Copyright 2017-2023 EPAM Systems, Inc. (https://www.epam.com/)
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+from pipeline.log.logger import PrintLogger
+
+
+class Logger:
+
+    inner = PrintLogger()
+
+    @staticmethod
+    def info(message, crucial=False, trace=False):
+        if crucial:
+            Logger.inner.info(message, trace=trace)
+        else:
+            Logger.inner.debug(message, trace=trace)
+
+    @staticmethod
+    def warn(message, crucial=False, trace=False):
+        if crucial:
+            Logger.inner.warning(message, trace=trace)
+        else:
+            Logger.inner.debug(message, trace=trace)
