+++ workflows/pipe-common/test/test_cpu_capacity_instance_selector.py
@@ -17,8 +17,8 @@ import logging
 import pytest
 from mock import MagicMock, Mock
 
-from pipeline.hpc.autoscaler import CpuCapacityInstanceSelector, IntegralDemand, InstanceDemand, Instance, \
-    FractionalDemand, ResourceSupply
+from pipeline.hpc.instance import CpuCapacityInstanceSelector, InstanceDemand, Instance
+from pipeline.hpc.resource import IntegralDemand, FractionalDemand, ResourceSupply
 
 logging.basicConfig(level=logging.DEBUG, format='%(asctime)s [%(threadName)s] [%(levelname)s] %(message)s')
