+++ workflows/pipe-common/test/test_scale_down.py
@@ -17,8 +17,10 @@ import logging
 from datetime import datetime, timedelta
 from mock import MagicMock, Mock
 
-from pipeline.hpc.autoscaler import GridEngineAutoscaler, GridEngineJob, GridEngineJobState, Clock, \
-    MemoryHostStorage
+from pipeline.hpc.autoscaler import GridEngineAutoscaler
+from pipeline.hpc.gridengine import GridEngineJob, GridEngineJobState
+from pipeline.hpc.host import MemoryHostStorage
+from pipeline.hpc.utils import Clock
 
 logging.basicConfig(level=logging.DEBUG, format='%(asctime)s [%(threadName)s] [%(levelname)s] %(message)s')
